exchange/requirements.txt
Django>=4.1,<5
django-cors-headers
djangorestframework-simplejwt
djangorestframework
drf-yasg  
drf-spectacular